# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/matching/property/{propertyId}": {
      "get": {
        "operationId": "MatchingController_getPropertyMatch",
        "summary": "Get match details for a specific property",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed match information for the property"
          }
        },
        "tags": [
          "Matching"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/properties/public/{id}": {
      "get": {
        "operationId": "PropertyController_findOnePublic",
        "summary": "Get a public property by ID (no auth required)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property found"
          },
          "404": {
            "description": "Property not found"
          }
        },
        "tags": [
          "properties"
        ]
      }
    },
...
    "/api/properties/upload/photos": {
      "post": {
        "operationId": "PropertyController_uploadPhotos",
        "summary": "Upload property photos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "photos": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Photos uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "url": {
                        "type": "string"
                      },
                      "key": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/properties/upload/video": {
      "post": {
        "operationId": "PropertyController_uploadVideo",
        "summary": "Upload property video",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "video": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Video uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/properties/upload/documents": {
      "post": {
        "operationId": "PropertyController_uploadDocuments",
        "summary": "Upload property documents",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "documents": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Documents uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/properties": {
      "post": {
        "operationId": "PropertyController_create",
        "summary": "Create a new property",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Property created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Building not found"
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PropertyController_findAll",
        "summary": "Get a page of properties",
        "parameters": [
          {
            "name": "bathrooms_min",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bathrooms",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bedrooms_min",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bedrooms",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "property_type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_landing_listing",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "operator_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "building_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Properties retrieved successfully"
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/properties/{id}": {
      "get": {
        "operationId": "PropertyController_findOne",
        "summary": "Get a property by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property found"
          },
          "404": {
            "description": "Property not found"
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "PropertyController_update",
        "summary": "Update a property",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Property not found"
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PropertyController_remove",
        "summary": "Delete a property",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property deleted successfully"
          },
          "404": {
            "description": "Property not found"
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TA-DA! Rental Platform API",
    "description": "API for connecting tenants and property operators",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer. Пример: 'Bearer 12345abcdef'",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
...
      "MetroStationDto": {
        "type": "object",
        "properties": {}
      },
      "CommuteTimeDto": {
        "type": "object",
        "properties": {}
      },
      "LocalEssentialDto": {
        "type": "object",
        "properties": {}
      },
      "PetDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePropertyDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Property title",
            "example": "Modern 2BR Apartment"
          },
          "building_id": {
            "type": "string",
            "description": "Building ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "operator_id": {
            "type": "string",
            "description": "Operator ID (required for private landlord properties)",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "apartment_number": {
            "type": "string",
            "description": "Apartment number",
            "example": "12A"
          },
          "descriptions": {
            "type": "string",
            "description": "Property descriptions",
            "example": "Beautiful modern apartment"
          },
          "property_type": {
            "type": "string",
            "description": "Property type",
            "example": "apartment",
            "enum": [
              "apartment",
              "flat",
              "studio",
              "penthouse",
              "room",
              "house",
              "maisonette",
              "duplex"
            ]
          },
          "furnishing": {
            "type": "string",
            "description": "Furnishing level",
            "example": "furnished",
            "enum": [
              "furnished",
              "unfurnished",
              "part_furnished"
            ]
          },
          "bills": {
            "type": "string",
            "description": "Bills included",
            "example": "included",
            "enum": [
              "included",
              "excluded"
            ]
          },
          "available_from": {
            "type": "string",
            "description": "Available from date",
            "example": "2024-01-15"
          },
          "building_type": {
            "type": "string",
            "description": "Building type",
            "example": "btr",
            "enum": [
              "btr",
              "co_living",
              "professional_management",
              "private_landlord"
            ]
          },
          "address": {
            "type": "string",
            "description": "Property address (inherited from building or custom)",
            "example": "123 Main St, London"
          },
          "tenant_types": {
            "description": "Tenant types for this property (inherited from building or custom)",
            "example": [
              "family",
              "student"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "amenities": {
            "description": "Property amenities (inherited from building or custom)",
            "example": [
              "Parking",
              "Garden",
              "Gym"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "property_amenities": {
            "description": "Apartment-level features (kitchen, bathroom, storage, tech, access)",
            "example": [
              "Dishwasher",
              "Rainfall shower",
              "Fibre broadband"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "family_status": {
            "type": "array",
            "description": "Target family statuses for this property (inherited from building or custom)",
            "example": [
              "couple",
              "couple-with-children"
            ],
            "items": {
              "type": "string",
              "enum": [
                "just-me",
                "couple",
                "couple-with-children",
                "single-parent",
                "friends-flatmates"
              ]
            }
          },
          "occupation": {
            "type": "array",
            "description": "Target occupations for this property (inherited from building or custom)",
            "example": [
              "student",
              "young-professional"
            ],
            "items": {
              "type": "string",
              "enum": [
                "student",
                "young-professional",
                "freelancer-remote-worker",
                "business-owner",
                "family-professional",
                "other"
              ]
            }
          },
          "children": {
            "type": "array",
            "description": "Target children statuses for this property (inherited from building or custom)",
            "example": [
              "no",
              "yes-1-child"
            ],
            "items": {
              "type": "string",
              "enum": [
                "no",
                "yes-1-child",
                "yes-2-children",
                "yes-3-plus-children"
              ]
            }
          },
          "pet_policy": {
            "type": "boolean",
            "description": "Pet policy (inherited from building or custom)",
            "example": true
          },
          "metro_stations": {
            "description": "Metro stations with travel times (inherited from building or custom)",
            "example": [
              {
                "label": "Oxford Circus",
                "destination": 5
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MetroStationDto"
            }
          },
          "commute_times": {
            "description": "Commute times to popular destinations (inherited from building or custom)",
            "example": [
              {
                "label": "City Centre",
                "destination": 15
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommuteTimeDto"
            }
          },
          "local_essentials": {
            "description": "Local essentials with distances (inherited from building or custom)",
            "example": [
              {
                "label": "Tesco Express",
                "destination": 200
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocalEssentialDto"
            }
          },
          "pets": {
            "description": "Allowed pet types and sizes (inherited from building or custom)",
            "example": [
              {
                "type": "dog",
                "size": "small"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PetDto"
            }
          },
          "let_duration": {
            "type": "string",
            "description": "Let duration — comma-separated multiselect of canonical tokens",
            "example": "12_months,long_term"
          },
          "floor": {
            "type": "number",
            "description": "Floor number",
            "example": 5
          },
          "square_meters": {
            "type": "number",
            "description": "Square meters",
            "example": 75.5
          },
          "balcony": {
            "type": "boolean",
            "description": "Has balcony",
            "example": true
          },
          "terrace": {
            "type": "boolean",
            "description": "Has terrace",
            "example": true
          },
          "price": {
            "type": "number",
            "description": "Monthly rent price (PCM)",
            "example": 2500
          },
          "deposit": {
            "type": "number",
            "description": "Security deposit",
            "example": 2500
          },
          "bedrooms": {
            "type": "number",
            "description": "Number of bedrooms",
            "example": 2
          },
          "bathrooms": {
            "type": "number",
            "description": "Number of bathrooms",
            "example": 2
          },
          "photos": {
            "description": "Property photos URLs",
            "example": [
              "https://s3.amazonaws.com/bucket/photo1.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "video": {
            "type": "string",
            "description": "Property video URL",
            "example": "https://s3.amazonaws.com/bucket/video.mp4"
          },
          "documents": {
            "type": "string",
            "description": "Property documents URL",
            "example": "https://s3.amazonaws.com/bucket/document.pdf"
          },
          "is_landing_listing": {
            "type": "boolean",
            "description": "Feature the property in the landing pages' listings section. Admin-only: the service drops this field for any other role.",
            "example": false
          },
          "postcode": {
            "type": "string",
            "description": "UK postcode. Geocoded via postcodes.io on save; when omitted, the full postcode is extracted from the address if present.",
            "example": "NW1 8XY"
          },
          "epc_rating": {
            "type": "string",
            "description": "EPC band (A-G). Legally required on listing advertisements in England and Wales.",
            "enum": [
              "A",
              "B",
              "C",
              "D",
              "E",
              "F",
              "G"
            ],
            "example": "C"
          },
          "status": {
            "type": "string",
            "description": "Listing lifecycle status. The booking pipeline drives listed -> under_offer -> let automatically; set it by hand to draft/archive a listing or re-list after a tenancy ends.",
            "enum": [
              "draft",
              "listed",
              "under_offer",
              "let",
              "archived"
            ],
            "example": "listed"
          }
        },
        "required": [
          "title"
        ]
      },
...
    }
  }
}