# backend/src/modules/property/property.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Patch,
  Param,
  ParseUUIDPipe,
  Delete,
  Query,
  UseInterceptors,
  UploadedFile,
  UploadedFiles,
  Logger,
  BadRequestException,
  InternalServerErrorException,
} from "@nestjs/common";
import { FileInterceptor, FilesInterceptor } from "@nestjs/platform-express";
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiBearerAuth,
  ApiConsumes,
  ApiBody,
  ApiQuery,
} from "@nestjs/swagger";
import { PropertyService } from "./property.service";
import { CreatePropertyDto } from "./dto/create-property.dto";
import { UpdatePropertyDto } from "./dto/update-property.dto";
import { FindPropertiesDto } from "./dto/find-properties.dto";
import { FindAdminPropertiesDto } from "./dto/find-admin-properties.dto";
import { Roles } from "@/common/decorators/roles.decorator";
import { UserRole } from "@/entities/user.entity";
import { Public } from "@/common/decorators/public.decorator";
import { S3Service } from "../../common/services/s3.service";
import { CurrentUser } from "../../common/decorators/current-user.decorator";
import { User } from "../../entities/user.entity";

@ApiTags("properties")
@Controller("properties")
export class PropertyController {
  private readonly logger = new Logger(PropertyController.name);

  constructor(
    private readonly propertyService: PropertyService,
    private readonly s3Service: S3Service,
  ) {}

  // Public endpoints - MUST be before protected routes
  @Get("public/all")
  @Public()
  @ApiOperation({ summary: "Get all public properties (no auth required)" })
  @ApiResponse({
    status: 200,
    description: "Public properties retrieved successfully",
  })
  async findAllPublic(@Query() query: FindPropertiesDto) {
    return this.propertyService.findAllPublic({
      page: query.page ? parseInt(query.page, 10) : undefined,
      limit: query.limit ? parseInt(query.limit, 10) : undefined,
      search: query.search,
      building_id: query.building_id,
    });
  }

  // Before `public/:id`: Nest matches routes in declaration order, and
  // "landing" would otherwise be read as an id.
  @Get("public/landing")
  @Public()
  @ApiOperation({
    summary: "Get the landing pages' featured listings (no auth required)",
  })
  @ApiResponse({
    status: 200,
    description: "Featured listings retrieved successfully",
  })
  async findLandingListings() {
    return this.propertyService.findLandingListings();
  }

  @Get("public/:id")
  @Public()
  @ApiOperation({ summary: "Get a public property by ID (no auth required)" })
  @ApiResponse({ status: 200, description: "Property found" })
  @ApiResponse({ status: 404, description: "Property not found" })
  async findOnePublic(@Param("id", ParseUUIDPipe) id: string) {
    return this.propertyService.findOnePublic(id);
  }

  @Get("public")
  @Public()
  @ApiOperation({
    summary: "Get paginated public properties (no auth required)",
  })
  @ApiResponse({
    status: 200,
    description: "Public properties retrieved successfully",
  })
  async getPublicProperties(@Query() query: FindPropertiesDto) {
    return this.propertyService.findAllPublic({
      page: query.page ? parseInt(query.page, 10) : undefined,
      limit: query.limit ? parseInt(query.limit, 10) : undefined,
      search: query.search,
      building_id: query.building_id,
    });
  }

  // Protected endpoints below
  @ApiBearerAuth()
  @Post("upload/photos")
  @Roles(UserRole.Admin, UserRole.Operator)
  @UseInterceptors(FilesInterceptor("photos", 1000))
  @ApiOperation({ summary: "Upload property photos" })
  @ApiConsumes("multipart/form-data")
  @ApiBody({
    schema: {
      type: "object",
      properties: {
        photos: {
          type: "array",
          items: {
            type: "string",
            format: "binary",
          },
        },
      },
    },
  })
  @ApiResponse({
    status: 201,
    description: "Photos uploaded successfully",
    schema: {
      type: "array",
      items: {
        type: "object",
        properties: {
          url: { type: "string" },
          key: { type: "string" },
        },
      },
    },
  })
  // Client mistakes (no file, wrong type) are 400s. As bare `Error` they came
  // back as 500s and paged Sentry for every wrong-file-type upload.
  async uploadPhotos(@UploadedFiles() files: Express.Multer.File[]) {
    if (!files || files.length === 0) {
      throw new BadRequestException("No files provided");
    }

    // Validated before any upload starts, so one bad file in a batch does not
    // leave the good ones already orphaned in S3.
    const invalid = files.find((file) => !file.mimetype.startsWith("image/"));
    if (invalid) {
      throw new BadRequestException(
        `Invalid file type for ${invalid.originalname}. Only images are allowed.`,
      );
    }

    try {
      return await Promise.all(
        files.map(async (file) => {
          const fileKey = this.s3Service.generateFileKey(
            file.originalname,
            "property-photo",
          );
          const uploadResult = await this.s3Service.uploadFile(
            file.buffer,
            fileKey,
            file.mimetype,
            file.originalname,
          );

          return {
            url: uploadResult.url,
            key: uploadResult.key,
          };
        }),
      );
    } catch (error) {
      this.logger.error("Photo upload failed", error?.stack ?? String(error));
      throw new InternalServerErrorException("Failed to upload photos");
    }
  }

  @ApiBearerAuth()
  @Post("upload/video")
  @Roles(UserRole.Admin, UserRole.Operator)
  @UseInterceptors(FileInterceptor("video"))
  @ApiOperation({ summary: "Upload property video" })
  @ApiConsumes("multipart/form-data")
  @ApiBody({
    schema: {
      type: "object",
      properties: {
        video: {
          type: "string",
          format: "binary",
        },
      },
    },
  })
  @ApiResponse({
    status: 201,
    description: "Video uploaded successfully",
    schema: {
      type: "object",
      properties: {
        url: { type: "string" },
        key: { type: "string" },
      },
    },
  })
  async uploadVideo(@UploadedFile() file: Express.Multer.File) {
    if (!file) {
      throw new BadRequestException("No file provided");
    }
    if (!file.mimetype.startsWith("video/")) {
      throw new BadRequestException("Invalid file type. Only videos are allowed.");
    }

    try {
      const fileKey = this.s3Service.generateFileKey(
        file.originalname,
        "property-video",
      );
      const uploadResult = await this.s3Service.uploadFile(
        file.buffer,
        fileKey,
        file.mimetype,
        file.originalname,
      );

      return {
        url: uploadResult.url,
        key: uploadResult.key,
      };
    } catch (error) {
      this.logger.error(`Error uploading video ${file.originalname}`, error?.stack ?? String(error));
      throw new InternalServerErrorException("Failed to upload video");
    }
  }

  @ApiBearerAuth()
  @Post("upload/documents")
  @Roles(UserRole.Admin, UserRole.Operator)
  @UseInterceptors(FileInterceptor("documents"))
  @ApiOperation({ summary: "Upload property documents" })
  @ApiConsumes("multipart/form-data")
  @ApiBody({
    schema: {
      type: "object",
      properties: {
        documents: {
          type: "string",
          format: "binary",
        },
      },
    },
  })
  @ApiResponse({
    status: 201,
    description: "Documents uploaded successfully",
    schema: {
      type: "object",
      properties: {
        url: { type: "string" },
        key: { type: "string" },
      },
    },
  })
  async uploadDocuments(@UploadedFile() file: Express.Multer.File) {
    if (!file) {
      throw new BadRequestException("No file provided");
    }
    if (file.mimetype !== "application/pdf") {
      throw new BadRequestException("Invalid file type. Only PDF files are allowed.");
    }

    try {
      const fileKey = this.s3Service.generateFileKey(
        file.originalname,
        "property-documents",
      );
      const uploadResult = await this.s3Service.uploadFile(
        file.buffer,
        fileKey,
        file.mimetype,
        file.originalname,
      );

      return {
        url: uploadResult.url,
        key: uploadResult.key,
      };
    } catch (error) {
      this.logger.error(`Error uploading document ${file.originalname}`, error?.stack ?? String(error));
      throw new InternalServerErrorException("Failed to upload document");
    }
  }

  // CRUD endpoints - after upload endpoints
  @ApiBearerAuth()
  @Post()
  @Roles(UserRole.Admin, UserRole.Operator)
  @ApiOperation({ summary: "Create a new property" })
  @ApiResponse({ status: 201, description: "Property created successfully" })
  @ApiResponse({ status: 400, description: "Bad request" })
  @ApiResponse({ status: 404, description: "Building not found" })
  create(
    @Body() createPropertyDto: CreatePropertyDto,
    @CurrentUser() user: User,
  ) {
    return this.propertyService.create(createPropertyDto, user.id, user.role);
  }

  @ApiBearerAuth()
  @Get()
  @Roles(UserRole.Admin, UserRole.Operator)
  @ApiOperation({ summary: "Get a page of properties" })
  @ApiQuery({ name: "page", required: false, schema: { type: "string" } })
  @ApiQuery({ name: "limit", required: false, schema: { type: "string" } })
  @ApiQuery({ name: "search", required: false, schema: { type: "string" } })
  @ApiQuery({
    name: "building_id",
    required: false,
    schema: { type: "string", format: "uuid" },
  })
  @ApiQuery({
    name: "operator_id",
    required: false,
    schema: { type: "string", format: "uuid" },
  })
  @ApiQuery({
    name: "is_landing_listing",
    required: false,
    schema: { type: "string", enum: ["true", "false"] },
  })
  @ApiQuery({
    name: "property_type",
    required: false,
    schema: { type: "string" },
  })
  @ApiQuery({ name: "bedrooms", required: false, schema: { type: "string" } })
  @ApiQuery({
    name: "bedrooms_min",
    required: false,
    schema: { type: "string" },
  })
  @ApiQuery({ name: "bathrooms", required: false, schema: { type: "string" } })
  @ApiQuery({
    name: "bathrooms_min",
    required: false,
    schema: { type: "string" },
  })
  @ApiResponse({
    status: 200,
    description: "Properties retrieved successfully",
  })
  async findAll(
    @CurrentUser() user: User,
    @Query() query: FindAdminPropertiesDto,
  ) {
    // Operators are always scoped to their own properties;
    // only admins may list everything or filter by an arbitrary operator.
    const operatorFilter =
      user.role === UserRole.Admin ? query.operator_id : user.id;
    return await this.propertyService.findAllWithFreshUrls({
      ...query,
      operator_id: operatorFilter,
    });
  }

  @ApiBearerAuth()
  @Get(":id")
  @Roles(UserRole.Admin, UserRole.Operator)
  @ApiOperation({ summary: "Get a property by ID" })
  @ApiResponse({ status: 200, description: "Property found" })
  @ApiResponse({ status: 404, description: "Property not found" })
  async findOne(@Param("id", ParseUUIDPipe) id: string, @CurrentUser() user: User) {
    return await this.propertyService.findOneWithFreshUrls(
      id,
      user.id,
      user.role,
    );
  }

  @ApiBearerAuth()
  @Patch(":id")
  @Roles(UserRole.Admin, UserRole.Operator)
  @ApiOperation({ summary: "Update a property" })
  @ApiResponse({ status: 200, description: "Property updated successfully" })
  @ApiResponse({ status: 400, description: "Bad request" })
  @ApiResponse({ status: 404, description: "Property not found" })
  update(
    @Param("id", ParseUUIDPipe) id: string,
    @Body() updatePropertyDto: UpdatePropertyDto,
    @CurrentUser() user: User,
  ) {
    return this.propertyService.update(
      id,
      updatePropertyDto,
      user.id,
      user.role,
    );
  }

  @ApiBearerAuth()
  @Delete(":id")
  @Roles(UserRole.Admin, UserRole.Operator)
  @ApiOperation({ summary: "Delete a property" })
  @ApiResponse({ status: 200, description: "Property deleted successfully" })
  @ApiResponse({ status: 404, description: "Property not found" })
  remove(@Param("id", ParseUUIDPipe) id: string, @CurrentUser() user: User) {
    return this.propertyService.remove(id, user.id, user.role);
  }
}